requirements.txt
# Scraping
scrapetube>=2.5.0
chromadb>=0.4.15
openai
python-dotenv
termcolor
sentence-transformers>=2.2.2

# LLM APIs
anthropic>=0.18.1
cohere>=4.37.0

# Utilities
python-dotenv>=1.0.0
PyYAML>=6.0
tqdm>=4.66.0
rank_bm25>=0.2.2
termcolor>=2.4.0
fastapi
uvicorn
pydantic
supabase
